# Speed up test_save_predictions