<!-- If you are submitting a dataset or a model for the model registry please use the corresponding checklists below otherwise feel free to remove them. -->
Fixes #1630

<!-- add additional description, question etc. related to the new dataset -->

Current slowest tests
```
============================= slowest 5 durations ==============================
20.86s call     tests/test_cli.py::test_run_task[intfloat/multilingual-e5-small-BornholmBitextMining-fd1525a9fd15316a2d503bf26ab031a61d056e98]
15.86s call     tests/test_cli.py::test_save_predictions
14.51s call     tests/test_tasks/test_mteb_rerank.py::test_mteb_rerank
12.34s call     tests/test_benchmark/test_benchmark_integration_with_datasets.py::test_benchmark_sentence_transformer[model0-BrazilianToxicTweetsClassification]
12.32s call     tests/test_benchmark/test_benchmark_integration_with_datasets.py::test_benchmark_sentence_transformer[model0-task1]
========= 1577 passed, 258 skipped, 274 warnings in 279.78s (0:04:39) ==========
```

## Checklist
<!-- Please do not delete this -->

- [x] Run tests locally to make sure nothing is broken using `make test`. 
- [x] Run the formatter to format the code using `make lint`.